Added header data checks.
Both adding and editing utilities now produce warnings when they detect
invalid data (i.e. missing information or forbidden characters). However,
unlike the similar function in the content building utility from which it
was adapted, the responsible function does not try to correct the erroneous
data, leaving it completely on the user.