Added basic colour support to blaze-list.
If the `color.list' option is set to `true' or `auto' (as in git), the
pages/posts listing is now displayed in fancy colours. I have also added a
couple of command-line options, `--color' and `--no-color', making it
possible to force either option no matter what is set in the configuration
file.