Removed uploading utility.
After a careful and long consideration, I have decided to move SKooDA's
blaze-submit utility to the separate package, blazeblogger-extras, instead
of distributing it as a part of core BlazeBlogger. Main reasons for this
are as follows:

  a) Unlike the core utilities that try to depend on default installation
     of Perl only, extra utilities tend to have additional dependencies,
     such as lftp. Separating these two is going to make life easier for
     both developers and packagers, as utility developers will have more
     freedom in choosing whatever additional tools suit them best, and
     packagers will know what exactly the additional dependencies are.

  b) Similarly to the previous point, while core utilities should remain
     platform independent, there is no reason to restrict extra utilities
     from taking advantage of the operating system.

  c) Releases of BlazeBlogger should not depend on the development of
     additional utilities and vice versa.

Nevertheless, support for these utilities -- and by this I mean namely
configuration options and presence in the command wrapper -- will of course
remain as it is.